Split module load & build to separate functions
